cortex-m: Get rid of basepri irq control
