+++ platform/opencv/templates/stm32f746g-discovery/mods.config
@@ -10,8 +10,7 @@ configuration conf {
 
 	//@Runlevel(0) include embox.arch.arm.fpu.cortex_m7_fp
 	//@Runlevel(0) include embox.arch.arm.fpu.fpv5(log_level=3)
-	@Runlevel(0) include embox.arch.arm.armmlib.interrupt(
-					nvic_use_prio=true, nvic_prio_shift=4)
+	@Runlevel(0) include embox.arch.arm.armmlib.interrupt
 	@Runlevel(1) include embox.driver.interrupt.cortexm_nvic(irq_table_size=128)
 	@Runlevel(1) include embox.driver.clock.cortexm_systick
 	@Runlevel(1) include embox.driver.serial.stm_usart_f7(baud_rate=115200, usartx=1)
