+++ src/arch/arm/armmlib/ipl_impl.h
@@ -10,23 +10,27 @@
 #endif /* HAL_IPL_H_ */
 
 #include <stdint.h>
-#include <framework/mod/options.h>
-
-#define NVIC_USE_PRIO \
-	OPTION_MODULE_GET(embox__arch__arm__armmlib__interrupt, \
-		BOOLEAN, nvic_use_prio)
-
-#define NVIC_PRIO_SHIFT \
-	OPTION_MODULE_GET(embox__arch__arm__armmlib__interrupt, \
-		NUMBER, nvic_prio_shift)
-
-#define NVIC_MIN_PRIO ((0xf << NVIC_PRIO_SHIFT) & 0xff)
-#define NVIC_MAX_PRIO ((0x1 << NVIC_PRIO_SHIFT) & 0xff)
 
 typedef uint32_t __ipl_t;
 
-#if NVIC_USE_PRIO
-#include "ipl_impl_prio.h"
-#else
-#include "ipl_impl_non_prio.h"
-#endif
+static inline void ipl_init(void) {
+	__asm__ __volatile__ (
+		"cpsie i \n\t");
+
+}
+
+static inline __ipl_t ipl_save(void) {
+	register uint32_t r;
+	__asm__ __volatile__ (
+		"mrs %0, PRIMASK;\n\t"
+		"cpsid i \n\t"
+		: "=r"(r));
+	return r;
+}
+
+static inline void ipl_restore(__ipl_t ipl) {
+	__asm__ __volatile__ (
+		"msr PRIMASK, %0;\n\t"
+		:
+		: "r"(ipl));
+}
