--- src/arch/arm/armmlib/ipl_impl_non_prio.h
@@ -1,33 +1,0 @@
-/**
- * @file
- *
- * @date 01.07.10
- * @author Anton Kozlov
- */
-
-#ifndef HAL_IPL_IMPL_H_
-#define HAL_IPL_IMPL_H_
-
-static inline void ipl_init(void) {
-	__asm__ __volatile__ (
-		"cpsie i \n\t");
-
-}
-
-static inline __ipl_t ipl_save(void) {
-	register uint32_t r;
-	__asm__ __volatile__ (
-		"mrs %0, PRIMASK;\n\t"
-		"cpsid i \n\t"
-		: "=r"(r));
-	return r;
-}
-
-static inline void ipl_restore(__ipl_t ipl) {
-	__asm__ __volatile__ (
-		"msr PRIMASK, %0;\n\t"
-		:
-		: "r"(ipl));
-}
-
-#endif /* HAL_IPL_IMPL_H_ */
