+++ templates/arm/stm32f746g-discovery-lua/mods.config
@@ -15,8 +15,7 @@ configuration conf {
 	/* Correctly works only with -O2 */
 	@Runlevel(2) include embox.test.math.fpu_context_consistency_test
 
-	@Runlevel(0) include embox.arch.arm.armmlib.interrupt(
-					nvic_use_prio=true, nvic_prio_shift=4)
+	@Runlevel(0) include embox.arch.arm.armmlib.interrupt
 
 	@Runlevel(1) include embox.driver.interrupt.cortexm_nvic(irq_table_size=128)
 	/*@Runlevel(1) include embox.driver.interrupt.cmsis_nvic*/
