+++ templates/arm/stm32f769i-discovery/mods.config
@@ -8,8 +8,7 @@ configuration conf {
 
 	@Runlevel(0) include embox.kernel.stack(stack_size=8192,alignment=4)
 
-	@Runlevel(0) include embox.arch.arm.armmlib.interrupt(
-					nvic_use_prio=true, nvic_prio_shift=4)
+	@Runlevel(0) include embox.arch.arm.armmlib.interrupt
 
 	@Runlevel(1) include embox.driver.interrupt.cortexm_nvic(irq_table_size=128)
 	@Runlevel(1) include embox.driver.clock.cortexm_systick
